backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyMuPDF==1.23.7
python-docx==1.1.2
transformers==4.39.3
torch
scikit-learn==1.6.1
PyPDF2==3.0.1
lxml==5.3.2
